Reorg test tangle builder/macros
