+++ bee-tangle/tests/common/mod.rs
@@ -1,0 +1,7 @@
+// Copyright 2021 IOTA Stiftung
+// SPDX-License-Identifier: Apache-2.0
+
+mod builder;
+mod macros;
+
+pub use builder::TangleBuilder;
